# chore: Tune Boogie encoding for better solver performance