--- a/Source/DafnyCore/Prelude/PreludeCore.bpl
+++ b/Source/DafnyCore/Prelude/PreludeCore.bpl
@@ -164,8 +164,8 @@ const $ArbitraryBoxValue: Box;
 
 function $Box<T>(T): Box;
 function $Unbox<T>(Box): T;
-axiom (forall<T> x : T   :: { $Box(x) } $Unbox($Box(x)) == x);
-axiom (forall<T> x : Box :: { $Unbox(x): T} $Box($Unbox(x): T) == x);
+axiom (forall<T> x : T   :: { $Box(x) } {:weight 3} $Unbox($Box(x)) == x);
+axiom (forall<T> x : Box :: { $Unbox(x): T}      $Box($Unbox(x): T) == x);
 
 
 // Corresponding entries for boxes...
@@ -462,7 +462,7 @@ axiom (forall o: ORDINAL, m,n: int ::
   { ORD#Plus(ORD#Plus(o, ORD#FromNat(m)), ORD#FromNat(n)) }
   0 <= m && 0 <= n ==>
   ORD#Plus(ORD#Plus(o, ORD#FromNat(m)), ORD#FromNat(n)) == ORD#Plus(o, ORD#FromNat(m+n)));
-// o-m-n == o+(m+n)
+// o-m-n == o-(m+n)
 axiom (forall o: ORDINAL, m,n: int ::
   { ORD#Minus(ORD#Minus(o, ORD#FromNat(m)), ORD#FromNat(n)) }
   0 <= m && 0 <= n && m+n <= ORD#Offset(o) ==>