--- a/Source/DafnyStandardLibraries/src/Std/Actions/Producers.dfy
+++ b/Source/DafnyStandardLibraries/src/Std/Actions/Producers.dfy
@@ -792,6 +792,7 @@ module Std.Producers {
         assert source.Outputs() == old(source.Outputs()) + sourceNewOutputs;
 
         assert Seq.All(Outputs(), IsSome);
+        assert old(!source.Done());
         OutputsPartitionedAfterOutputtingSome(result.value);
         ProduceSome(result.value);
         assert (Seq.All(source.Outputs(), IsSome) ==> Seq.All(Outputs(), IsSome));