--- a/Source/IntegrationTests/LitTests.cs
+++ b/Source/IntegrationTests/LitTests.cs
@@ -33,7 +33,7 @@ namespace IntegrationTests {
       "/proverOpt:O:auto_config=false",
       "/proverOpt:O:type_check=true",
       "/proverOpt:O:smt.case_split=3",
-      "/proverOpt:O:smt.qi.eager_threshold=100",
+      "/proverOpt:O:smt.qi.eager_threshold=44",
       "/proverOpt:O:smt.delay_units=true",
       "/proverOpt:O:smt.arith.solver=2",
       "/proverOpt:PROVER_PATH:" + RepositoryRoot + $"../unzippedRelease/dafny/z3/bin/z3-{DafnyOptions.DefaultZ3Version}"