--- a/Source/IntegrationTests/TestFiles/LitTests/LitTest/dafny1/SchorrWaite.dfy.expect
+++ b/Source/IntegrationTests/TestFiles/LitTests/LitTest/dafny1/SchorrWaite.dfy.expect
@@ -1,4 +1,4 @@
 
 Dafny program verifier finished with 272 verified, 0 errors
-Total resources used is 27749890
-Max resources used by VC is 2181911
+Total resources used is 27490314
+Max resources used by VC is 2376436