--- a/Source/IntegrationTests/TestFiles/LitTests/LitTest/dafny1/SchorrWaite.dfy.refresh.expect
+++ b/Source/IntegrationTests/TestFiles/LitTests/LitTest/dafny1/SchorrWaite.dfy.refresh.expect
@@ -1,4 +1,4 @@
 
 Dafny program verifier finished with 276 verified, 0 errors
-Total resources used is 28898941
-Max resources used by VC is 1253781
+Total resources used is 26163152
+Max resources used by VC is 1684695