… Source/DafnyCore/GeneratedFromDafny/Defs.cs
… Source/DafnyPipeline.Test/expectedProverLog.smt2
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-cs.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-go.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-java.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-js.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-notarget.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries-py.doo
… Source/DafnyStandardLibraries/binaries/DafnyStandardLibraries.doo