This PR contains three changes to our Boogie encoding to improve performance of Z3 when solving Dafny generated problems:
- A higher weight for an axiom defining `$Unbox` and `$Box` as each other's inverse
- Lower `eager_threshold` and linear scaling of existing weights
- A more restrictive trigger on the generated axioms defining the rank order on inductive datatypes

I arrived at those by benchmarking on a set of code bases, most of which are closed source. They have been pseudonymized (A.0, A.1, B.old, and B.new). A.0 and A.1 share some code and talk about the same problem. B.old and B.new are snapshots of the same code base, but changes have been made to improve proof stability. The open source one, are the Dafny standard libraries. To ensure the changes don't overfit, I added another code base (C) on the final evaluation and tested on a more recent version of Z3 (4.14.1 vs. 4.12.1). The table below shows the changes to the runtime of Z3 on those code bases when making the proposed changes.

| Z3 | libraries | A.0 | A.1 | B.old | B.new | C |
| ------ | ------ | ------ | ------ | ------ | ------ | ------ |
| 4.12.1  | -29,58 % | -18,86 % | -16,67 % | -13,06 % | -3,54 % | -18,83 % |
| 4.14.1  | -10,25 % | -58,84 % | -38,46 % | -17,97 % | -4,01 % | -14,41 % |

These results convince me, that the proposed changes, will also help all our users now and going forward. Even though we see proofs breaking, for all codebases, they are outweighed by the ones that start working. Below, there are more detailed statistics.

### Z3 v4.12.1
```
Comparing libraries...
Analyzed 4027 queries. Of those, 4019 lined up and 2 stopped working, while 6 started working.
The overall runtime improved by 29.58 %.
Of the matching queries 1163 (28.94 %, 59.57 % runtime share) significantly improved, 12 (0.30 %, 0.71 % runtime share) experienced a marked slowdown, and 2844 (70.76 %) stayed roughly the same.
Comparing A.0...
Analyzed 574 queries. Of those, 572 lined up and 1 stopped working, while 1 started working.
The overall runtime improved by 18.87 %.
Of the matching queries 214 (37.41 %, 39.24 % runtime share) significantly improved, 6 (1.05 %, 18.12 % runtime share) experienced a marked slowdown, and 352 (61.54 %) stayed roughly the same.
Comparing A.1...
Analyzed 1207 queries. Of those, 1205 lined up and 0 stopped working, while 2 started working.
The overall runtime improved by 16.67 %.
Of the matching queries 158 (13.11 %, 44.57 % runtime share) significantly improved, 12 (1.00 %, 3.52 % runtime share) experienced a marked slowdown, and 1035 (85.89 %) stayed roughly the same.
Comparing B.old...
Analyzed 12691 queries. Of those, 12688 lined up and 1 stopped working, while 2 started working.
The overall runtime improved by 13.06 %.
Of the matching queries 93 (0.73 %, 12.94 % runtime share) significantly improved, 11 (0.09 %, 0.58 % runtime share) experienced a marked slowdown, and 12584 (99.18 %) stayed roughly the same.
Comparing B.new...
Analyzed 18009 queries. Of those, 18006 lined up and 1 stopped working, while 2 started working.
The overall runtime improved by 3.54 %.
Of the matching queries 89 (0.49 %, 4.19 % runtime share) significantly improved, 52 (0.29 %, 0.45 % runtime share) experienced a marked slowdown, and 17865 (99.22 %) stayed roughly the same.
Comparing C...
Analyzed 25268 queries. Of those, 25249 lined up and 7 stopped working, while 12 started working.
The overall runtime improved by 18.83 %.
Of the matching queries 8699 (34.45 %, 39.91 % runtime share) significantly improved, 23 (0.09 %, 0.95 % runtime share) experienced a marked slowdown, and 16527 (65.46 %) stayed roughly the same.
```

### Z3 v4.14.1
```
Comparing libraries...
Analyzed 4014 queries. Of those, 4001 lined up and 6 stopped working, while 7 started working.
The overall runtime improved by 10.25 %.
Of the matching queries 1722 (43.04 %, 64.31 % runtime share) significantly improved, 31 (0.77 %, 5.40 % runtime share) experienced a marked slowdown, and 2248 (56.19 %) stayed roughly the same.
Comparing A.0...
Analyzed 582 queries. Of those, 576 lined up and 1 stopped working, while 5 started working.
The overall runtime improved by 58.84 %.
Of the matching queries 263 (45.66 %, 77.02 % runtime share) significantly improved, 7 (1.22 %, 5.66 % runtime share) experienced a marked slowdown, and 306 (53.12 %) stayed roughly the same.
Comparing A.1...
Analyzed 1211 queries. Of those, 1204 lined up and 5 stopped working, while 2 started working.
The overall runtime improved by 38.46 %.
Of the matching queries 391 (32.48 %, 61.11 % runtime share) significantly improved, 14 (1.16 %, 1.27 % runtime share) experienced a marked slowdown, and 799 (66.36 %) stayed roughly the same.
Comparing B.old...
Analyzed 12682 queries. Of those, 12665 lined up and 14 stopped working, while 3 started working.
The overall runtime improved by 17.97 %.
Of the matching queries 200 (1.58 %, 21.59 % runtime share) significantly improved, 25 (0.20 %, 0.90 % runtime share) experienced a marked slowdown, and 12440 (98.22 %) stayed roughly the same.
Comparing B.new...
Analyzed 18013 queries. Of those, 18000 lined up and 10 stopped working, while 3 started working.
The overall runtime improved by 4.01 %.
Of the matching queries 338 (1.88 %, 12.96 % runtime share) significantly improved, 45 (0.25 %, 0.64 % runtime share) experienced a marked slowdown, and 17617 (97.87 %) stayed roughly the same.
Comparing C...
Analyzed 25249 queries. Of those, 25233 lined up and 7 stopped working, while 9 started working.
The overall runtime improved by 14.41 %.
Of the matching queries 10444 (41.39 %, 44.85 % runtime share) significantly improved, 54 (0.21 %, 2.87 % runtime share) experienced a marked slowdown, and 14735 (58.40 %) stayed roughly the same.
```